hassio-to-pubsub-source/cloud-functions/hass-to-bigquery/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery==1.8.1
